use NoDial option for opening streams in non active relays
It turns out that the relays are doing dialing on NewStream in the wild,
presumably because they still have dead connections.
This uses the new NoDial option to instruct the swarm to not dial the hop target
instead of explicitly checking for open connections and is robust to failed
connections.