feat: remove relay discovery and unspecified relay dialing
This was a misfeature that looks great in demos but isn't useful in practice.

First, random relays we discover are unlikely to support _active_ relaying. This
means that picking a random relay is usually going to fail because the random
relay won't already be connected to the target.

Second, there are two use-cases for relaying:

1. Connecting to undialable nodes.
2. Relaying all outbound traffic.

For the first case, we only want to use the relays specified by the target
peers. For the second case, we'd need to modify either libp2p's dialer logic (in
the swarm) to prefix all addresses with the specified relay.

The logic _here_ covers neither use-case.